Add files via upload
Use the plotly.express library to create a scatter plot. Visualize the relationship between the sepal length and sepal width. Adjust the appearance of the chart to give a different color to each of the species.
Create an app object based on the Dash class.
Create an application template consisting of:
H1 header with the content: "Hello dash"
H3 header with the content: "Dash: A web application framework for Python."
a Graph object to which you will pass the chart created in step 5. Pass any object identifier to the argument named id.

Based on the data in the df variable, create a unique list of species values.
Modify the content of the H3 header - Choose species:.
In the application template, under the H1 and H3 headings, define another application element - a drop-down list:
- give an identifier to the Dropdown object so that it can be referenced in a callback
- define a default (choose it yourself) list value.
- pass the previously created list of values to the corresponding argument
- set the appropriate parameter to prevent the selected value from being cleared
Create a callback to change the chart depending on the value of the species parameter.
- in the function, load and filter the dataset by the parameter selected from the drop-down list
- return a scatter plot of the relationship between the sepal length and sepal width.